Integrates remaining [fitted]util and [fitted]util_tests
